refactor: moves core logic into @app-config/core
