+++ app-config-core/.eslintrc.js
@@ -1,0 +1,1 @@
+module.exports = require('@lcdev/eslint-config/cwd')(__dirname);
