+++ app-config-core/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@app-config/core",
+  "description": "Core logic for @lcdev/app-config",
+  "version": "2.0.0-rc.9",
+  "license": "MPL-2.0",
+  "author": {
+    "name": "Launchcode",
+    "email": "admin@lc.dev",
+    "url": "https://lc.dev"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/launchcodedev/app-config.git"
+  },
+  "main": "dist/index.js",
+  "module": "dist/es/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "/dist",
+    "!*.tsbuildinfo",
+    "!*.test.*"
+  ],
+  "scripts": {
+    "build": "tsc -b",
+    "build:es": "tsc -b tsconfig.es.json",
+    "clean": "rm -rf dist *.tsbuildinfo",
+    "lint": "eslint src",
+    "fix": "eslint --fix src",
+    "test": "jest",
+    "prepublishOnly": "yarn clean && yarn build && yarn build:es"
+  },
+  "dependencies": {
+    "@iarna/toml": "3",
+    "js-yaml": "3",
+    "json5": "2",
+    "lodash.merge": "4",
+    "util": "0.12"
+  },
+  "devDependencies": {
+    "@types/js-yaml": "3",
+    "@types/lodash.merge": "4"
+  },
+  "prettier": "@lcdev/prettier",
+  "jest": {
+    "preset": "@lcdev/jest"
+  }
+}
