+++ app-config-core/src/errors.ts
