+++ app-config/src/cli.ts
@@ -7,14 +7,19 @@ import { outputFile, readFile } from 'fs-extra';
 import { resolve } from 'json-schema-ref-parser';
 import { stripIndents } from 'common-tags';
 import {
-  promptUser,
-  consumeStdin,
-  flattenObjectTree,
-  renameInFlattenedTree,
+  FileType,
   Json,
   JsonObject,
-} from './common';
-import { FileType, stringify } from './config-source';
+  stringify,
+  flattenObjectTree,
+  renameInFlattenedTree,
+  logger,
+  LogLevel,
+  AppConfigError,
+  FailedToSelectSubObject,
+  EmptyStdinOrPromptResponse,
+} from '@app-config/core';
+import { checkTTY, promptUser, consumeStdin } from './common';
 import {
   Configuration,
   Options as LoadConfigOptions,
@@ -43,8 +48,6 @@ import { shouldUseSecretAgent, startAgent, disconnectAgents } from './secret-age
 import { loadSchema } from './schema';
 import { generateTypeFiles } from './generate';
 import { validateAllConfigVariants } from './validation';
-import { checkTTY, logger, LogLevel } from './logging';
-import { AppConfigError, FailedToSelectSubObject, EmptyStdinOrPromptResponse } from './errors';
 
 enum OptionGroups {
   Options = 'Options:',
