+++ app-config/src/config-source.test.ts
@@ -1,77 +1,5 @@
-import {
-  CombinedSource,
-  EnvironmentSource,
-  FallbackSource,
-  FileSource,
-  FlexibleFileSource,
-  LiteralSource,
-} from './config-source';
-import { ParsingExtension } from './parsed-value';
+import { EnvironmentSource, FileSource, FlexibleFileSource } from './config-source';
 import { withTempFiles } from './test-util';
-
-const flattenExtension: ParsingExtension = (value, [_, key]) => {
-  if (key === '$flatten') {
-    return (parse) => parse(value, { shouldFlatten: true });
-  }
-
-  return false;
-};
-
-const uppercaseExtension: ParsingExtension = (value) => {
-  if (typeof value === 'string') {
-    return (parse) => parse(value.toUpperCase());
-  }
-
-  return false;
-};
-
-describe('Parsing', () => {
-  it('interprets an empty object', async () => {
-    const source = new LiteralSource({});
-    const parsed = await source.read();
-
-    expect(parsed.sources[0]).toBe(source);
-    expect(parsed.raw).toEqual({});
-    expect(parsed.toJSON()).toEqual({});
-    expect(parsed.toString()).toEqual('{}');
-  });
-
-  it('uses parsing extensions', async () => {
-    const parsed = await new LiteralSource({
-      $flatten: 'bar',
-    }).read([flattenExtension]);
-
-    expect(parsed.toJSON()).toEqual('bar');
-  });
-
-  it('uses parsing extensions in nested objects', async () => {
-    const parsed = await new LiteralSource({
-      foo: {
-        $flatten: 'bar',
-      },
-    }).read([flattenExtension]);
-
-    expect(parsed.toJSON()).toEqual({ foo: 'bar' });
-  });
-
-  it('uses value transform extension', async () => {
-    const parsed = await new LiteralSource({
-      foo: 'bar',
-    }).read([uppercaseExtension]);
-
-    expect(parsed.toJSON()).toEqual({ foo: 'BAR' });
-  });
-
-  it('uses multiple extensions', async () => {
-    const parsed = await new LiteralSource({
-      foo: {
-        $flatten: 'bar',
-      },
-    }).read([flattenExtension, uppercaseExtension]);
-
-    expect(parsed.toJSON()).toEqual({ foo: 'BAR' });
-  });
-});
 
 describe('FileSource', () => {
   it('fails when file doesnt exist', async () => {
@@ -248,56 +176,5 @@ describe('EnvironmentSource', () => {
 
     expect(parsed.sources[0]).toBe(source);
     expect(parsed.toJSON()).toEqual({ foo: 'bar' });
-  });
-});
-
-describe('CombinedSource', () => {
-  it('fails when no sources are provided', () => {
-    expect(() => new CombinedSource([])).toThrow();
-  });
-
-  it('combines a few sources', async () => {
-    const source = new CombinedSource([
-      new LiteralSource({ foo: 1 }),
-      new LiteralSource({ bar: 2 }),
-      new LiteralSource({ baz: 3 }),
-    ]);
-
-    const parsed = await source.read();
-
-    expect(parsed.sources[0]).toBe(source);
-    expect(parsed.toJSON()).toEqual({ foo: 1, bar: 2, baz: 3 });
-  });
-});
-
-describe('FallbackSource', () => {
-  it('fails when no sources are provided', () => {
-    expect(() => new FallbackSource([])).toThrow();
-  });
-
-  it('selects the first source', async () => {
-    const source = new FallbackSource([
-      new LiteralSource({ foo: 1 }),
-      new LiteralSource({ bar: 2 }),
-      new LiteralSource({ baz: 3 }),
-    ]);
-
-    const parsed = await source.read();
-
-    expect(parsed.sources[0]).toBe(source.sources[0]);
-    expect(parsed.toJSON()).toEqual({ foo: 1 });
-  });
-
-  it('selects source when first one fails', async () => {
-    const source = new FallbackSource([
-      new FileSource('test-file.json'),
-      new LiteralSource({ bar: 2 }),
-      new LiteralSource({ baz: 3 }),
-    ]);
-
-    const parsed = await source.read();
-
-    expect(parsed.sources[0]).toBe(source.sources[1]);
-    expect(parsed.toJSON()).toEqual({ bar: 2 });
   });
 });
