+++ app-config/src/secret-agent.ts
@@ -1,7 +1,7 @@
 import https from 'https';
 import WebSocket from 'ws';
 import { Server as BaseServer, Client as BaseClient, MessageVariant } from '@lcdev/ws-rpc/bson';
-import { Json } from './common';
+import { Json, AppConfigError, logger } from '@app-config/core';
 import {
   Key,
   decryptValue,
@@ -13,8 +13,6 @@ import {
 } from './encryption';
 import { loadOrCreateCert } from './secret-agent-tls';
 import { loadSettingsLazy, saveSettings } from './settings';
-import { AppConfigError } from './errors';
-import { logger } from './logging';
 
 export enum MessageType {
   Ping = 'Ping',
